+++ src/plugins/fsl_native/noise_sampler/noise_sampler_sample.c
@@ -1,6 +1,28 @@
-#include "deps/fossil/math/noise.h"
+/*!
+ *  Copyright 2026 Lily Awertnex
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
 
-#include "noise.h"
+/*!
+ *  @file noise_sampler_sample.c
+ *
+ *  @brief general noise functions used to parse samples.
+ */
+
+#include "../../../math/noise.h"
+
+#include "noise_sampler_sample.h"
 
 #include <math.h>
 
@@ -25,18 +47,18 @@
 #define RAND_CONST_12 904023
 #define RAND_CONST_13 371769
 
-f64 nolerp_f64(const f64 *n, const f64 *t)
+f64 fsl_noise_sample_nolerp(const f64 *n, const f64 *t)
 {
     (void)t;
     return n[0];
 }
 
-f64 lerp_f64(const f64 *n, const f64 *t)
+f64 fsl_noise_sample_lerp(const f64 *n, const f64 *t)
 {
     return n[0] + (n[1] - n[0]) * t[0];
 }
 
-f64 bilerp_f64(const f64 *n, const f64 *t)
+f64 fsl_noise_sample_bilerp(const f64 *n, const f64 *t)
 {
     f64 w[2] = {0};
     w[0] = 1.0 - t[0];
@@ -48,7 +70,7 @@ f64 bilerp_f64(const f64 *n, const f64 *t)
         n[3] * t[0] * t[1];
 }
 
-f64 trilerp_f64(const f64 *n, const f64 *t)
+f64 fsl_noise_sample_trilerp(const f64 *n, const f64 *t)
 {
     f64 w[3] = {0};
     w[0] = 1.0 - t[0];
@@ -65,7 +87,7 @@ f64 trilerp_f64(const f64 *n, const f64 *t)
         n[7] * t[0] * t[1] * t[2];
 }
 
-v2f64 gradient_2d(i32 x, i32 y, u64 seed)
+v2f64 fsl_noise_sample_gradient_2d(i32 x, i32 y, u64 seed)
 {
     v2f64 v = {0};
     u64 h = {0};
@@ -79,7 +101,7 @@ v2f64 gradient_2d(i32 x, i32 y, u64 seed)
     return v;
 }
 
-v3f64 gradient_3d(i32 x, i32 y, i32 z, u64 seed)
+v3f64 fsl_noise_sample_gradient_3d(i32 x, i32 y, i32 z, u64 seed)
 {
     v3f64 v = {0};
     u64 h = {0};
@@ -94,7 +116,7 @@ v3f64 gradient_3d(i32 x, i32 y, i32 z, u64 seed)
     return v;
 }
 
-chunk_work_cost noise_sample_axis_init(hhc_noise_sample *s, u8 axis, f64 pos, f64 frequency)
+void fsl_noise_sample_axis_init(fsl_noise_sample *s, u8 axis, f64 pos, f64 frequency)
 {
     f64 v = pos * frequency;
     i64 a = (i64)floorf(v);
@@ -109,10 +131,9 @@ chunk_work_cost noise_sample_axis_init(hhc_noise_sample *s, u8 axis, f64 pos, f6
     s->dw[axis] = 1.0 - d;
     s->da[axis] = v - (f64)a;
     s->db[axis] = v - (f64)b;
-    return CHUNK_WORK_COST_GENERATE_NOISE_INIT;
 }
 
-chunk_work_cost noise_sample_make_2d(const hhc_noise_sample *s, f64 *dst, f64 amplitude, u64 seed)
+f64 fsl_noise_sample_make_2d(const fsl_noise_sample *s, f64 amplitude, u64 seed)
 {
     const f64 dx = s->dv[0];
     const f64 dy = s->dv[1];
@@ -121,25 +142,24 @@ chunk_work_cost noise_sample_make_2d(const hhc_noise_sample *s, f64 *dst, f64 am
     f64 n[4] = {0};
     v2f64 g[4] = {0};
 
-    g[0] = gradient_2d(s->a[0], s->a[1], seed);
-    g[1] = gradient_2d(s->b[0], s->a[1], seed);
-    g[2] = gradient_2d(s->a[0], s->b[1], seed);
-    g[3] = gradient_2d(s->b[0], s->b[1], seed);
+    g[0] = fsl_noise_sample_gradient_2d(s->a[0], s->a[1], seed);
+    g[1] = fsl_noise_sample_gradient_2d(s->b[0], s->a[1], seed);
+    g[2] = fsl_noise_sample_gradient_2d(s->a[0], s->b[1], seed);
+    g[3] = fsl_noise_sample_gradient_2d(s->b[0], s->b[1], seed);
 
     n[0] = s->da[0] * g[0].x + s->da[1] * g[0].y;
     n[1] = s->db[0] * g[1].x + s->da[1] * g[1].y;
     n[2] = s->da[0] * g[2].x + s->db[1] * g[2].y;
     n[3] = s->db[0] * g[3].x + s->db[1] * g[3].y;
 
-    *dst =
+    return
         (n[0] * wx * wy +
          n[1] * dx * wy +
          n[2] * wx * dy +
          n[3] * dx * dy) * amplitude;
-    return CHUNK_WORK_COST_GENERATE_NOISE_INTERPOLATE_2D;
 }
 
-chunk_work_cost noise_sample_make_3d(const hhc_noise_sample *s, f64 *dst, f64 amplitude, u64 seed)
+f64 fsl_noise_sample_make_3d(const fsl_noise_sample *s, f64 amplitude, u64 seed)
 {
     const f64 dx = s->dv[0];
     const f64 dy = s->dv[1];
@@ -150,14 +170,14 @@ chunk_work_cost noise_sample_make_3d(const hhc_noise_sample *s, f64 *dst, f64 am
     f64 n[8] = {0};
     v3f64 g[8] = {0};
 
-    g[0] = gradient_3d(s->a[0], s->a[1], s->a[2], seed);
-    g[1] = gradient_3d(s->b[0], s->a[1], s->a[2], seed);
-    g[2] = gradient_3d(s->a[0], s->b[1], s->a[2], seed);
-    g[3] = gradient_3d(s->b[0], s->b[1], s->a[2], seed);
-    g[4] = gradient_3d(s->a[0], s->a[1], s->b[2], seed);
-    g[5] = gradient_3d(s->b[0], s->a[1], s->b[2], seed);
-    g[6] = gradient_3d(s->a[0], s->b[1], s->b[2], seed);
-    g[7] = gradient_3d(s->b[0], s->b[1], s->b[2], seed);
+    g[0] = fsl_noise_sample_gradient_3d(s->a[0], s->a[1], s->a[2], seed);
+    g[1] = fsl_noise_sample_gradient_3d(s->b[0], s->a[1], s->a[2], seed);
+    g[2] = fsl_noise_sample_gradient_3d(s->a[0], s->b[1], s->a[2], seed);
+    g[3] = fsl_noise_sample_gradient_3d(s->b[0], s->b[1], s->a[2], seed);
+    g[4] = fsl_noise_sample_gradient_3d(s->a[0], s->a[1], s->b[2], seed);
+    g[5] = fsl_noise_sample_gradient_3d(s->b[0], s->a[1], s->b[2], seed);
+    g[6] = fsl_noise_sample_gradient_3d(s->a[0], s->b[1], s->b[2], seed);
+    g[7] = fsl_noise_sample_gradient_3d(s->b[0], s->b[1], s->b[2], seed);
 
     n[0] = s->da[0] * g[0].x + s->da[1] * g[0].y + s->da[2] * g[0].z;
     n[1] = s->db[0] * g[1].x + s->da[1] * g[1].y + s->da[2] * g[1].z;
@@ -168,7 +188,7 @@ chunk_work_cost noise_sample_make_3d(const hhc_noise_sample *s, f64 *dst, f64 am
     n[6] = s->da[0] * g[6].x + s->db[1] * g[6].y + s->db[2] * g[6].z;
     n[7] = s->db[0] * g[7].x + s->db[1] * g[7].y + s->db[2] * g[7].z;
 
-    *dst =
+    return
         (n[0] * wx * wy * wz +
          n[1] * dx * wy * wz +
          n[2] * wx * dy * wz +
@@ -177,5 +197,4 @@ chunk_work_cost noise_sample_make_3d(const hhc_noise_sample *s, f64 *dst, f64 am
          n[5] * dx * wy * dz +
          n[6] * wx * dy * dz +
          n[7] * dx * dy * dz) * amplitude;
-    return CHUNK_WORK_COST_GENERATE_NOISE_INTERPOLATE_3D;
 }